Refactor Postgres timeseries objects
This change moves the SortedTimeseries and NormalizedTimeseries
pg_type macros to a new extension/time_series.rs file.
